# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Solve Corporate CRM API
  version: 1.0.0
  description: API do CRM Solve Corporate - Ecossistema Digital Integrado
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all plans
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Plan"
                  total:
                    type: integer
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price]
              properties:
                name:
                  type: string
                  minLength: 2
                description:
                  type: string
                price:
                  type: integer
                  minimum: 0
                periodicity:
                  type: string
                  enum: [mensal, trimestral, semestral, anual]
                  default: mensal
                duration:
                  type: integer
                active:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get plan by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Plan details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Plan"
        "404":
          description: Plan not found
    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Plan"
        "404":
          description: Plan not found
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete plan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Plan deleted
        "404":
          description: Plan not found
  /plans/{id}/toggle:
    patch:
      operationId: togglePlan
      tags: [plans]
      summary: Toggle plan active status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Plan toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Plan"
        "404":
          description: Plan not found
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    type: integer
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Deactivate user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deactivated
        "404":
          description: User not found
  /users/{id}/toggle:
    patch:
      operationId: toggleUser
      tags: [users]
      summary: Toggle user active status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "404":
          description: User not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [administrador, gestor, comercial, financeiro, operacional]
        phone:
          type: string
        active:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        role:
          type: string
          enum: [administrador, gestor, comercial, financeiro, operacional]
        phone:
          type: string
        active:
          type: boolean
...
    Plan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        periodicity:
          type: string
          enum: [mensal, trimestral, semestral, anual]
        duration:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlanUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
        price:
          type: integer
          minimum: 0
        periodicity:
          type: string
          enum: [mensal, trimestral, semestral, anual]
        duration:
          type: integer
        active:
          type: boolean
...